Widen BlockingStatus.Timer to float64 for Pi-hole v6
Pi-hole v6 returns a float for the blocking timer during a temporary-disable countdown, so decoding into an int fails intermittently.

+++ internal/pihole/model.go
@@ -4,7 +4,7 @@ import "fmt"
 
 type BlockingStatus struct {
 	Blocking string  `json:"blocking"`
-	Timer    int     `json:"timer"`
+	Timer    float64 `json:"timer"`
 	Took     float64 `json:"took"`
 }
